# IADocumentClassifier.API/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "IA Document Classifier API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Permissions/{id}": {
      "get": {
        "tags": [
          "Permissions"
        ],
        "summary": "Metodo para consultar por permisos",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Permissions"
        ],
        "summary": "Metodo para Eliminar permisos",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Roles/{id}": {
      "get": {
        "tags": [
          "Roles"
        ],
        "summary": "Metodo para consultar por Roles",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Roles"
        ],
        "summary": "Metodo para Eliminar o borrar Roles",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/RolesPermission": {
      "get": {
        "tags": [
          "RolesPermission"
        ],
        "summary": "Metodo para consultar todos los roles por permisos",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "RolesPermission"
        ],
        "summary": "Metodo para crear Roles permisos",
        "requestBody": {
          "description": "",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/RolesPermissionDTO"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RolesPermissionDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RolesPermissionDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RolesPermissionDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "RolesPermission"
        ],
        "summary": "Metodo para actualizar roles permisos",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "",
            "schema": {
              "type": "integer",
              "description": "",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/RolesPermissionDTO"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RolesPermissionDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RolesPermissionDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RolesPermissionDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/RolesPermission/{id}": {
      "get": {
        "tags": [
          "RolesPermission"
        ],
        "summary": "Metodo para consultar por roles por permisos",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "RolesPermission"
        ],
        "summary": "Metodo para borrar roles permisos",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RolesPermissionDTO": {
        "type": "object",
        "properties": {
          "rolPermission_Id": {
            "type": "integer",
            "format": "int32"
          },
          "rol_Id": {
            "type": "integer",
            "format": "int32"
          },
          "permission_Id": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}